# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CruiseBooking",
        "title": "CruiseBooking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/cabin-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CabinType"
                ],
                "summary": "List cabin types by cruise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cruise ID",
                        "name": "cruise_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CabinType"
                ],
                "summary": "Create a cabin type",
                "parameters": [
                    {
                        "description": "Cabin type data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CabinTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/cabin-types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CabinType"
                ],
                "summary": "Update a cabin type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CabinType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cabin type data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CabinTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "CabinType"
                ],
                "summary": "Delete a cabin type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CabinType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cruisebooking_backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码（0 表示成功）",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.CabinTypeRequest": {
            "type": "object",
            "required": [
                "cruise_id",
                "name"
            ],
            "properties": {
                "amenities": {
                    "description": "设施清单",
                    "type": "string"
                },
                "area": {
                    "description": "面积（平方米）",
                    "type": "number"
                },
                "area_max": {
                    "description": "最大面积",
                    "type": "number"
                },
                "area_min": {
                    "description": "最小面积",
                    "type": "number"
                },
                "bed_type": {
                    "description": "床型说明",
                    "type": "string"
                },
                "capacity": {
                    "description": "容纳人数",
                    "type": "integer"
                },
                "code": {
                    "description": "舱房类型代码",
                    "type": "string"
                },
                "cruise_id": {
                    "description": "所属邮轮 ID",
                    "type": "integer"
                },
                "deck": {
                    "description": "所在甲板",
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "english_name": {
                    "description": "英文名称",
                    "type": "string"
                },
                "floor_plan_url": {
                    "description": "平面图 URL",
                    "type": "string"
                },
                "max_capacity": {
                    "description": "最大容纳人数",
                    "type": "integer"
                },
                "name": {
                    "description": "舱房类型名称",
                    "type": "string"
                },
                "sort_order": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "tags": {
                    "description": "特色标签",
                    "type": "string"
                }
            }
        },
...
    }
}